mv BridgeWebViewClient to single file
