+++ library/src/main/java/com/github/lzyzsd/jsbridge/BridgeWebView.java
@@ -79,7 +79,7 @@ public class BridgeWebView extends WebView implements WebViewJavascriptBridge {
         return new BridgeWebViewClient(this);
     }
 
-	private void handlerReturnData(String url) {
+	void handlerReturnData(String url) {
 		String functionName = BridgeUtil.getFunctionFromReturnUrl(url);
 		CallBackFunction f = responseCallbacks.get(functionName);
 		String data = BridgeUtil.getDataFromReturnUrl(url);
@@ -89,59 +89,6 @@ public class BridgeWebView extends WebView implements WebViewJavascriptBridge {
 			return;
 		}
 	}
-
-	public static class BridgeWebViewClient extends WebViewClient {
-		private BridgeWebView webView;
-
-		public BridgeWebViewClient(BridgeWebView webView) {
-			this.webView = webView;
-		}
-
-        @Override
-		public boolean shouldOverrideUrlLoading(WebView view, String url) {
-			try {
-				url = URLDecoder.decode(url, "UTF-8");
-			} catch (UnsupportedEncodingException e) {
-				e.printStackTrace();
-			}
-			if (url.startsWith(BridgeUtil.YY_RETURN_DATA)) { // 如果是返回数据
-				webView.handlerReturnData(url);
-				return true;
-			} else if (url.startsWith(BridgeUtil.YY_OVERRIDE_SCHEMA)) { //
-				webView.flushMessageQueue();
-				return true;
-			} else {
-				return super.shouldOverrideUrlLoading(view, url);
-			}
-		}
-
-		@Override
-		public void onPageStarted(WebView view, String url, Bitmap favicon) {
-			super.onPageStarted(view, url, favicon);
-		}
-
-		@Override
-		public void onPageFinished(WebView view, String url) {
-			super.onPageFinished(view, url);
-
-			if (BridgeWebView.toLoadJs != null) {
-				BridgeUtil.webViewLoadLocalJs(view, BridgeWebView.toLoadJs);
-			}
-
-			//
-			if (webView.getStartupMessage() != null) {
-				for (Message m : webView.getStartupMessage()) {
-					webView.dispatchMessage(m);
-				}
-				webView.setStartupMessage(null);
-			}
-		}
-
-        @Override
-        public void onReceivedError(WebView view, int errorCode, String description, String failingUrl) {
-            super.onReceivedError(view, errorCode, description, failingUrl);
-        }
-    }
 
 	@Override
 	public void send(String data) {
@@ -177,7 +124,7 @@ public class BridgeWebView extends WebView implements WebViewJavascriptBridge {
 		}
 	}
 
-	private void dispatchMessage(Message m) {
+	void dispatchMessage(Message m) {
         String messageJson = m.toJson();
         //escape special characters for json string
         messageJson = messageJson.replaceAll("(\\\\)([^utrn])", "\\\\\\\\$1$2");
@@ -188,7 +135,7 @@ public class BridgeWebView extends WebView implements WebViewJavascriptBridge {
         }
     }
 
-	public void flushMessageQueue() {
+	void flushMessageQueue() {
 		if (Thread.currentThread() == Looper.getMainLooper().getThread()) {
 			loadUrl(BridgeUtil.JS_FETCH_QUEUE_FROM_JAVA, new CallBackFunction() {
